Py/backend/requirements.txt
flask
flask-cors
pymongo
dnspython
geopy
python-dotenv
